Separate Control elements from the Canvas class
Logic for game control elements like buttons, dropdowns, and other widgets have
been seperated (as much as possible -- with the exception of the slider)
from the `Canvas` class, who's main responsibility is handling just the
canvas
