+++ canvas.py
@@ -35,66 +35,6 @@ class GameCanvas:
         )
         self.canvas.pack(anchor=tk.CENTER, expand=True, side=tk.TOP)
 
-        # Main frame for control buttons
-        control_frame = tk.Frame(self.root)
-        control_frame.pack(padx=50, side=tk.RIGHT, expand=True)
-
-        # first subframe for choosing patterns
-        pattern_control = tk.Frame(control_frame)
-        pattern_control.pack(pady=10, side='top')
-
-        # large spacer between top and bottom button groups
-        spacer = tk.Frame(control_frame)
-        spacer.pack(pady=50, expand=True)
-
-        # second subframe for controlling canvas simulation
-        sim_control = tk.Frame(control_frame)
-        sim_control.pack(pady=10, side='top')
-
-        # Buttons in pattern control
-        self.random_btn = tk.Button(
-            # button for random patterns
-            pattern_control,
-            text="Random",
-            command=self.randomize_grid,
-            **button_style,
-        )
-
-        # Pattern selection dropdown
-        self.pattern_var = tk.StringVar(self.root)
-        self.pattern_var.set("Select Pattern")  # Default value
-        self.pattern_menu = tk.OptionMenu(
-            pattern_control,
-            self.pattern_var,
-            *PREDEFINED_PATTERNS.keys(),
-            command=self.load_pattern,
-        )
-        self.pattern_menu.config(**button_style)  # Style dropdown
-        self.pattern_menu.pack(side=tk.TOP, padx=5, expand=True)
-
-        # buttons in sim control
-        self.start_btn = tk.Button(
-            # button for starting the simulation
-            sim_control,
-            text="Start",
-            command=self.start_game,
-            **button_style,
-        ).pack(side=tk.TOP, padx=5)
-
-        self.stop_btn = tk.Button(
-            sim_control, text="Stop", command=self.stop_game, **button_style
-        )
-        self.stop_btn.pack(side=tk.TOP, padx=5)
-
-        self.random_btn.pack(side=tk.TOP, padx=5)
-        self.clear_btn = tk.Button(
-            sim_control,
-            text="Clear",
-            command=self.clear_grid,
-            **button_style,
-        )
-        self.clear_btn.pack(side=tk.TOP, padx=5)
-
         # Speed control slider
         self.speed_slider = tk.Scale(
             canvas_slider,
@@ -112,33 +52,11 @@ class GameCanvas:
         self.speed_slider.pack(side=tk.TOP, pady=5, expand=True)
 
         # add custom speed label under the slider
-        speed_text = tk.Label(canvas_slider, text='speed', font=font_style.get('default'))
+        speed_text = tk.Label(canvas_slider, text='speed (ms)', font=font_style.get('default'))
         speed_text.pack(side=tk.TOP, expand=True)
 
         # Bind mouse click to toggle cells
         self.canvas.bind("<Button-1>", self.game.toggle_cell)
-
-    def start_game(self):
-        """Starts updating the game state"""
-        self.game.start()
-
-    def stop_game(self):
-        """Stops updating the game state"""
-        self.game.stop()
-
-    def randomize_grid(self):
-        """Randomizes the placement of dead and alive cells on the canvas"""
-        self.game.randomize_grid()
-        self.render_canvas()
-
-    def clear_grid(self):
-        """"""
-        self.game.clear_grid()
-        self.render_canvas()
-
-    def load_pattern(self, pattern_name):
-        self.game.load_pattern(pattern_name)
-        self.render_canvas()
 
     def set_speed(self, value):
         self.game.set_speed(value)
